feat: reset bonding curve to canonical 18-decimal integer protocol (#1922)
* feat(bonding-curve): reset testnet protocol to canonical integer model

* fix(zhtp): align bonding curve handlers with u128 core

* fix(lib-client): widen bonding curve ffi builder calls

* fix(zhtp-cli): widen curve command transaction amounts

* Implement canonical bonding curve math foundation

* fix(bonding-curve): address review issues from PR #1922

- Fix reserve/treasury split from 40/60 to 20/80 per canonical spec
- Assign remainder to reserve (ceiling 20%) not treasury for exact accounting
- Fix oracle staleness: preserve graduation_pending_since_block on stale price
  to prevent confirmation window reset exploit
- Fix PiecewiseLinear u128→u64 clamping in price_at and quote_buy: return 0
  instead of silently clamping to u64::MAX with 18-decimal inputs
- Fix MINIMUM_AMM_LIQUIDITY: 1_000_000 → 10_000_000_000_000_000 (0.01 SOV
  at 18 decimals, not 10^-12 SOV)
- Fix test scale: use TOKEN_SCALE instead of legacy 100_000_000 factor

* fix(bonding-curve): revert 40/60 split to correct value per issue #1844

The Copilot review incorrectly claimed the spec defined 20/80. The
codebase consistently documents 40% reserve / 60% treasury throughout
comments, tests, and inline notes. Restore the correct split.

Also preserve the remainder-to-reserve (ceiling 40%) arithmetic fix
from the previous commit to avoid destroying atomic units per tx.

+++ lib-blockchain/src/transaction/core.rs
@@ -2404,9 +2404,9 @@ pub struct BondingCurveBuyData {
     /// Token ID being purchased
     pub token_id: [u8; 32],
     /// Amount of stablecoin to spend
-    pub stable_amount: u64,
+    pub stable_amount: u128,
     /// Minimum tokens expected (slippage protection)
-    pub min_tokens_out: u64,
+    pub min_tokens_out: u128,
     /// Buyer's address
     pub buyer: [u8; 32],
     /// Nonce for replay protection
@@ -2419,9 +2419,9 @@ pub struct BondingCurveSellData {
     /// Token ID being sold
     pub token_id: [u8; 32],
     /// Amount of tokens to sell
-    pub token_amount: u64,
+    pub token_amount: u128,
     /// Minimum stablecoin expected (slippage protection)
-    pub min_stable_out: u64,
+    pub min_stable_out: u128,
     /// Seller's address
     pub seller: [u8; 32],
     /// Nonce for replay protection
@@ -2436,9 +2436,9 @@ pub struct BondingCurveGraduateData {
     /// AMM pool ID to create/use
     pub pool_id: [u8; 32],
     /// SOV amount to seed into AMM pool
-    pub sov_seed_amount: u64,
+    pub sov_seed_amount: u128,
     /// Token amount to seed into AMM pool
-    pub token_seed_amount: u64,
+    pub token_seed_amount: u128,
     /// Graduator's address (must be creator or governance)
     pub graduator: [u8; 32],
     /// Nonce for replay protection
